+++ source/geh_calculated_measurements/src/geh_calculated_measurements/electrical_heating/application/calculation.py
@@ -2,7 +2,7 @@ from geh_common.domain.types import MeteringPointType, OrchestrationType
 from geh_common.telemetry.decorators import use_span
 from pyspark.sql import SparkSession
 
-from geh_calculated_measurements.common.domain.model import calculated_measurements_factory
+from geh_calculated_measurements.common.application.model import calculated_measurements_factory
 from geh_calculated_measurements.common.infrastructure import (
     CalculatedMeasurementsRepository,
     CurrentMeasurementsRepository,
@@ -35,6 +35,7 @@ def execute_application(spark: SparkSession, args: ElectricalHeatingArgs) -> Non
         consumption_metering_point_periods,
         child_metering_point_periods,
         args.time_zone,
+        args.execution_start_datetime,
     )
 
     # Write the calculated measurements
@@ -44,6 +45,7 @@ def execute_application(spark: SparkSession, args: ElectricalHeatingArgs) -> Non
         orchestration_type=OrchestrationType.ELECTRICAL_HEATING,
         metering_point_type=MeteringPointType.ELECTRICAL_HEATING,
         time_zone=args.time_zone,
+        transaction_creation_datetime=args.execution_start_datetime,
     )
     calculated_measurements_repository = CalculatedMeasurementsRepository(spark, args.catalog_name)
     calculated_measurements_repository.write_calculated_measurements(calculated_measurements)